add props basic example
